pyproject.toml
[tool.poetry]
name = "django-oemof"
version = "0.1.0"
description = "Django application to run oemof"
authors = ["Hendrik Huyskens <hendrik.huyskens@rl-institut.de>"]
license = "AGPL-3.0"
readme = "README.md"
packages = [{include = "django_oemof"}]

[tool.poetry.dependencies]
python = "^3.9"
django-environ = "^0.4.5"  # https://github.com/joke2k/django-environ
oemoflex = {git = "https://github.com/rl-institut/oemoflex.git", rev = "dev"}


[tool.poetry.group.dev.dependencies]
pre-commit = "^2.20.0"
black = "^21.12b0"
flake8 = "^5.0.4"
pylint = "^2.15.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.black]
line-length = 120

[tool.isort]
profile = "black"
